requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.23
coreapi==2.0.9
Django==1.10.1
django-multiselectfield==0.1.7
django-solo==1.1.2
django-widget-tweaks==1.4.1
djangorestframework==3.5.1
fuzzywuzzy==0.15.1
kombu==3.0.35
Pillow>=3.5,<=4.0
PyPDF2==1.26.0
python-Levenshtein==0.12.0
rarfile==2.8
requests==2.11.1
requests-cache==0.4.12
